refactored styling
