+++ src/styles/closingDialog.css
