+++ src/styles/openingDialog.css
